+++ src/client/ClientGameRunner.ts
@@ -1127,7 +1127,7 @@ export class ClientGameRunner {
           const closestOfType = this.gameView
             .nearbyUnits(
               clickedTile,
-              this.gameView.config().openBackSnapRadius(bu.type),
+              this.gameView.config().structureMinDist(),
               bu.type,
             )
             .filter(({ unit }) => unit.owner().id() === myPlayerID)